Move build step from build.rs to main.rs in ibc-prost-compiler
