+++ ibc_prost_compiler/Cargo.toml
@@ -4,11 +4,7 @@ version = "0.1.0"
 authors = ["Greg Szabo <greg@philosobear.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 walkdir = { version = "2.3.1" }
-
-[build-dependencies]
 prost-build = { version = "0.6.1" }
-walkdir = { version = "2.3.1" }
+tempdir = "0.3.7"
